* commit '561f8821e976cfcdd5503e1b645417ef6506a995':
  Prevent possible NullPointerException when service is destroyed

+++ src/com/android/bluetooth/btservice/ProfileService.java
@@ -174,7 +174,8 @@ public abstract class ProfileService extends Service {
     @Override
     public void onDestroy() {
         if (DBG) log("Destroying service.");
-        mAdapterService.removeProfile(this);
+        if (mAdapterService != null) mAdapterService.removeProfile(this);
+
         if (mCleaningUp) {
             if (DBG) log("Cleanup already started... Skipping cleanup()...");
         } else {
